+++ example/use_cases/employer_university_flow/employer_university_flow/trust.go
@@ -1,0 +1,16 @@
+package employer_university_flow
+
+type trustedEntitiesStore struct {
+	Issuers map[string]bool
+}
+
+func (t *trustedEntitiesStore) isTrusted(did string) bool {
+	if v, ok := t.Issuers[did]; ok {
+		return v
+	}
+	return false
+}
+
+var TrustedEntities = trustedEntitiesStore{
+	Issuers: make(map[string]bool),
+}
